// frontend/src/types/visitorProfileData.ts
export interface WeddingDetailsData {
    firstName: string;
    lastName: string;
    partnerFirstName: string;
    partnerLastName: string;
    engagementDate: string;
    weddingDate: string;
    weddingVenue: string;
}

export interface AccountDetailsData {
  email: string;
  password: string;
  retypePassword: string;
}